refactor: centralize backend constants
