+++ backend/src/features/users/models/users/admin/admin.model.js
@@ -1,5 +1,5 @@
 import User from "../user.model.js";
-import {  Roles  } from '../../../../../../../docs/constants/roles.js';
+import { Roles } from '../../../../../../../shared/constants/index.js';
 
 class Admin extends User {
     static role = Roles.ADMIN;
