+++ backend/src/shared/constants/index.js
@@ -1,6 +1,8 @@
 import HTTP_STATUS from "./http-codes.js";
-import { Roles } from "../../../../docs/constants/roles.js";
-import { Status } from '../../../../docs/constants/status.js';
+import { Roles } from "./roles.js";
+import { Status } from "./status.js";
+import { Disciplines } from "./disciplines.js";
+import { Events } from "./events.js";
 
 const Actions = Object.freeze({
     DESK_REJECT: 'desk-reject',
@@ -27,4 +29,4 @@ const Actions = Object.freeze({
     PUBLISHED_BY_EIC: 'published-by-eic',
 });
 
-export { Roles, HTTP_STATUS, Status, Actions };
+export { Roles, HTTP_STATUS, Status, Disciplines, Events, Actions };
