+++ backend/src/shared/utils/generate-id.js
@@ -1,5 +1,4 @@
 import { nanoid } from 'nanoid';
-// const { Status } = require('../../../../docs/constants/status.js');
 
 
 /*------------------------------------------------------------------------------*/
